+++ src/app/user/__init__.py
@@ -1,0 +1,4 @@
+__author__ = 'djw'
+from .forms import SettingsForm
+from .models import User, update_user, create_user
+from .views import get_authed_user, settings, user_dashboard
