+++ src/app/user/models.py
@@ -1,5 +1,5 @@
 from google.appengine.ext import ndb
-from models import BaseModel
+from app.models.base import BaseModel
 
 
 class User(BaseModel):
